+++ src/api/informer.rs
@@ -4,7 +4,7 @@ use crate::api::{
     Void,
 };
 use crate::api::resource::{
-    ResourceList,
+    ObjectList,
     WatchEvent,
 };
 use crate::client::APIClient;
@@ -159,7 +159,7 @@ impl<P, U> Informer<P, U> where
         let req = self.resource.list_zero_resource_entries(&self.params)?;
 
         // parse to void a ResourceList into void except for Metadata
-        let res = self.client.request::<ResourceList<Void>>(req)?;
+        let res = self.client.request::<ObjectList<Void>>(req)?;
 
         let version = res.metadata.resourceVersion.unwrap_or_else(|| "0".into());
         debug!("Got fresh resourceVersion={} for {}", version, self.resource.resource);